requirements.txt
pycryptodome
msgpack
python-dotenv
aiohttp
asyncio
beautifulsoup4
SQLAlchemy
apscheduler
google-play-scraper